requirements.txt
yfinance
pandas
numpy
plotly
dash
dash-bootstrap-components
scikit-learn
python-dotenv
loguru
